Refactor Timestamp into Container & Component
