+++ package.json
@@ -58,6 +58,7 @@
   },
   "homepage": "https://github.com/bbc-news/simorgh#readme",
   "dependencies": {
+    "@bbc/gel-foundations": "^0.1.3",
     "@bbc/psammead-brand": "^0.1.1",
     "@bbc/psammead-caption": "^0.1.0",
     "@bbc/psammead-copyright": "^0.2.0",
