+++ src/app/containers/Timestamp/index.stories.jsx
@@ -2,6 +2,6 @@ import React from 'react';
 import { storiesOf } from '@storybook/react'; // eslint-disable-line import/no-extraneous-dependencies
 import Timestamp from './index';
 
-storiesOf('Timestamp', module).add('default', () => (
-  <Timestamp datetime="1530947227000" datetimeText="7 July 2018" />
+storiesOf('TimestampContainer', module).add('default', () => (
+  <Timestamp timestamp={1530947227000} />
 ));
